Move builder classes into builder.py
