+++ guppylang-internals/src/guppylang_internals/compiler/cfg_compiler.py
@@ -14,10 +14,9 @@ from guppylang_internals.checker.cfg_checker import (
     Signature,
 )
 from guppylang_internals.checker.core import Place, Variable
+from guppylang_internals.compiler.builder import BlockBuilder, DFBuilder
 from guppylang_internals.compiler.core import (
-    BlockBuilder,
     CompilerContext,
-    DFBuilder,
     DFContainer,
     is_return_var,
     return_var,
