+++ guppylang-internals/src/guppylang_internals/definition/custom.py
@@ -20,11 +20,10 @@ from guppylang_internals.ast_util import (
 from guppylang_internals.checker.core import Context, Globals
 from guppylang_internals.checker.expr_checker import check_call, synthesize_call
 from guppylang_internals.checker.func_checker import check_signature
+from guppylang_internals.compiler.builder import DFBuilder, FunctionBuilder
 from guppylang_internals.compiler.core import (
     CompilerContext,
-    DFBuilder,
     DFContainer,
-    FunctionBuilder,
     GlobalConstId,
 )
 from guppylang_internals.definition.common import CheckableGenericDef, ParsableDef
